Make the hashCode64 and updateHash features of the Value type rather than the AnyValue type.
