+++ community/values/src/main/java/org/neo4j/values/storable/Value.java
@@ -28,6 +28,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.neo4j.graphdb.spatial.Geometry;
+import org.neo4j.hashing.HashFunction;
 import org.neo4j.values.AnyValue;
 import org.neo4j.values.AnyValueWriter;
 import org.neo4j.values.Comparison;
@@ -228,6 +229,15 @@ public abstract class Value extends AnyValue
     public abstract ValueGroup valueGroup();
 
     public abstract NumberType numberType();
+
+    public final long hashCode64()
+    {
+        HashFunction xxh64 = HashFunction.incrementalXXH64();
+        long seed = 1; // Arbitrary seed, but it must always be the same or hash values will change.
+        return xxh64.finalise( updateHash( xxh64, xxh64.initialise( seed ) ) );
+    }
+
+    public abstract long updateHash( HashFunction hashFunction, long hash );
 
     public boolean isNaN()
     {
